refactor(open-mpm): split cli/memories_cmd + search_cmd into submodules (#366)
